docker-deploy/web-app/requirements.txt
Django>3.0
psycopg2
djangorestframework
django-cors-headers
